+++ obadiah/src/main/java/net/zionsoft/obadiah/biblereading/NfcHelper.java
@@ -28,6 +28,8 @@ import android.os.Build;
 import android.support.annotation.Nullable;
 import android.support.v4.content.ContextCompat;
 
+import net.zionsoft.obadiah.utils.UriBuilder;
+
 class NfcHelper {
     static void registerNdefMessageCallback(Activity activity,
                                             NfcAdapter.CreateNdefMessageCallback callback) {
@@ -46,7 +48,7 @@ class NfcHelper {
     static NdefMessage createNdefMessage(Context context, String translation,
                                          int book, int chapter, int verse) {
         return new NdefMessage(new NdefRecord[]{
-                NdefRecord.createUri(UriHelper.createUri(translation, book, chapter, verse)),
+                NdefRecord.createUri(UriBuilder.createUri(translation, book, chapter, verse)),
                 NdefRecord.createApplicationRecord(context.getPackageName())});
     }
 }