+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 #### Next Release
 - Added setting & analytics for daily verse.
 - Added link to Facebook page.
+- Added notes to Firebase App Indexing.
 
 #### v1.16.3 (2016-12-26)
 - Fixed cursor jumping in notes due to synchronization.